undo explicit bfloat16, move to main

+++ graphgps/encoder/laplace_pos_encoder.py
@@ -44,8 +44,6 @@ class LapPENodeEncoder(torch.nn.Module):
 
         # Initial projection of eigenvalue and the node's eigenvector value
         self.linear_A = nn.Linear(2, dim_pe)
-        # Convert the entire linear_A layer to bfloat16
-        self.linear_A = self.linear_A.to(torch.bfloat16)
 
         if norm_type == 'batchnorm':
             self.raw_norm = nn.BatchNorm1d(max_freqs)
@@ -98,8 +96,8 @@ class LapPENodeEncoder(torch.nn.Module):
                              f"required for {self.__class__.__name__}; "
                              "set config 'posenc_LapPE.enable' to True")
             
-        EigVals = batch.EigVals.to(torch.bfloat16)
-        EigVecs = batch.EigVecs.to(torch.bfloat16)
+        EigVals = batch.EigVals
+        EigVecs = batch.EigVecs
         
         if self.training:
             sign_flip = torch.rand(EigVecs.size(1), device=EigVecs.device)
